+++ apps/prairielearn/src/models/enrollment.ts
@@ -11,7 +11,7 @@ import {
 import { run } from '@prairielearn/run';
 
 import {
-  PotentialEnterpriseEnrollmentStatus,
+  PotentialEnrollmentStatus,
   checkPotentialEnterpriseEnrollment,
 } from '../ee/models/enrollment.js';
 import {
@@ -107,7 +107,7 @@ async function _enrollUserInCourseInstance({
 }): Promise<Enrollment> {
   assertHasRole(authzData, requiredRole);
 
-  assertEnrollmentStatus(lockedEnrollment, ['invited', 'removed', 'rejected']);
+  assertEnrollmentStatus(lockedEnrollment, ['invited', 'left', 'removed', 'rejected']);
 
   const newEnrollment = await queryRow(
     sql.enroll_user,
@@ -137,7 +137,7 @@ async function _enrollUserInCourseInstance({
  * enrollment already exists, this is a no-op. This function does not check
  * enterprise enrollment eligibility, and should not be used directly outside of tests.
  *
- * If the user was in the 'removed', 'invited' or 'rejected' status, this will set the
+ * If the user was in the 'left', 'removed', 'invited' or 'rejected' status, this will set the
  * enrollment status to 'joined'.
  *
  * If the user was 'blocked', this will throw an error.
@@ -235,6 +235,7 @@ export async function ensureEnrollment({
   authzData,
   requiredRole,
   actionDetail,
+  throwOnIneligible = true,
 }: {
   institution: Institution;
   course: Course;
@@ -242,13 +243,16 @@ export async function ensureEnrollment({
   authzData: Exclude<AuthzDataWithoutEffectiveUser, DangerousSystemAuthzData>;
   requiredRole: 'Student'[];
   actionDetail: SupportedActionsForTable<'enrollments'>;
-}) {
+  throwOnIneligible?: boolean;
+}): Promise<PotentialEnrollmentStatus> {
   // If the current user is not a student, bail.
   // We don't want to give instructors an enrollment.
-  if (!hasRole(authzData, requiredRole)) return;
+  if (!hasRole(authzData, requiredRole)) return PotentialEnrollmentStatus.INELIGIBLE;
+
+  let status = PotentialEnrollmentStatus.ALLOWED;
 
   if (isEnterprise()) {
-    const status = await checkPotentialEnterpriseEnrollment({
+    status = await checkPotentialEnterpriseEnrollment({
       institution,
       course,
       courseInstance,
@@ -256,11 +260,21 @@ export async function ensureEnrollment({
     });
 
     switch (status) {
-      case PotentialEnterpriseEnrollmentStatus.PLAN_GRANTS_REQUIRED:
-        throw new HttpRedirect(`/pl/course_instance/${courseInstance.id}/upgrade`);
-      case PotentialEnterpriseEnrollmentStatus.LIMIT_EXCEEDED:
-        throw new HttpRedirect('/pl/enroll/limit_exceeded');
-      case PotentialEnterpriseEnrollmentStatus.ALLOWED:
+      case PotentialEnrollmentStatus.PLAN_GRANTS_REQUIRED: {
+        if (throwOnIneligible) {
+          throw new HttpRedirect(`/pl/course_instance/${courseInstance.id}/upgrade`);
+        } else {
+          return status;
+        }
+      }
+      case PotentialEnrollmentStatus.LIMIT_EXCEEDED: {
+        if (throwOnIneligible) {
+          throw new HttpRedirect('/pl/enroll/limit_exceeded');
+        } else {
+          return status;
+        }
+      }
+      case PotentialEnrollmentStatus.ALLOWED:
         break;
       default:
         assertNever(status);
@@ -274,6 +288,8 @@ export async function ensureEnrollment({
     authzData,
     actionDetail,
   });
+
+  return status;
 }
 
 export async function selectOptionalEnrollmentByUserId({
@@ -441,10 +457,11 @@ async function _inviteExistingEnrollment({
   lockedEnrollment: Enrollment;
   pendingUid: string;
   authzData: AuthzDataWithEffectiveUser;
-  requiredRole: 'Student Data Editor'[];
+  requiredRole: ('Student Data Editor' | 'System')[];
 }): Promise<Enrollment> {
   assertHasRole(authzData, requiredRole);
-  assertEnrollmentStatus(lockedEnrollment, ['rejected', 'removed', 'blocked']);
+  // TODO: remove the `removed` status in https://github.com/PrairieLearn/PrairieLearn/pull/13803.
+  assertEnrollmentStatus(lockedEnrollment, ['rejected', 'left', 'removed', 'blocked']);
 
   const newEnrollment = await queryRow(
     sql.invite_existing_enrollment,
@@ -476,7 +493,7 @@ async function inviteNewEnrollment({
   pendingUid: string;
   authzData: AuthzDataWithEffectiveUser;
   courseInstance: CourseInstanceContext;
-  requiredRole: 'Student Data Editor'[];
+  requiredRole: ('Student Data Editor' | 'System')[];
 }) {
   assertHasRole(authzData, requiredRole);
   const newEnrollment = await queryRow(
@@ -504,7 +521,7 @@ async function inviteNewEnrollment({
  * If there is an existing enrollment with the given uid, it will be updated to a invitation.
  * If there is no existing enrollment, a new enrollment will be created.
  *
- * Transitions users in the 'blocked', 'rejected' or 'removed' status to 'invited'.
+ * Transitions users in the 'blocked', 'rejected', 'left', or 'removed' status to 'invited'.
  */
 export async function inviteStudentByUid({
   uid,
@@ -513,7 +530,7 @@ export async function inviteStudentByUid({
   requiredRole,
 }: {
   uid: string;
-  requiredRole: 'Student Data Editor'[];
+  requiredRole: ('Student Data Editor' | 'System')[];
   authzData: AuthzDataWithEffectiveUser;
   courseInstance: CourseInstanceContext;
 }): Promise<Enrollment> {
@@ -565,11 +582,12 @@ export async function setEnrollmentStatus({
   requiredRole,
 }: {
   enrollment: Enrollment;
-  status: 'rejected' | 'blocked' | 'removed' | 'joined';
+  status: 'rejected' | 'blocked' | 'left' | 'removed' | 'joined';
   authzData: AuthzData;
   requiredRole: CourseInstanceRole[];
 }): Promise<Enrollment> {
   const transitionInformation: {
+    equivalentStatuses?: EnumEnrollmentStatus[];
     previousStatus: EnumEnrollmentStatus;
     actionDetail: SupportedActionsForTable<'enrollments'>;
     permittedRoles: CourseInstanceRole[];
@@ -581,11 +599,20 @@ export async function setEnrollmentStatus({
           actionDetail: 'unblocked',
           permittedRoles: ['Student Data Viewer', 'Student Data Editor'],
         };
+      case 'left':
+        return {
+          // If a student tries to leave a course but has already been removed by
+          // an instructor, we will treat this as a no-op.
+          equivalentStatuses: ['removed'],
+          previousStatus: 'joined',
+          actionDetail: 'left',
+          permittedRoles: ['Student'],
+        };
       case 'removed':
         return {
           previousStatus: 'joined',
           actionDetail: 'removed',
-          permittedRoles: ['Student'],
+          permittedRoles: ['Student Data Viewer', 'Student Data Editor'],
         };
       case 'rejected':
         return {
@@ -609,8 +636,12 @@ export async function setEnrollmentStatus({
     if (lockedEnrollment.user_id) {
       await selectAndLockUser(lockedEnrollment.user_id);
     }
-    // The enrollment is already in the desired status, so we can return early.
-    if (lockedEnrollment.status === status) {
+
+    if (
+      lockedEnrollment.status === status ||
+      transitionInformation.equivalentStatuses?.includes(lockedEnrollment.status)
+    ) {
+      // The enrollment is already in the desired status, so we can return early.
       return lockedEnrollment;
     }