+++ apps/prairielearn/src/pages/instructorInstanceAdminSettings/components/SelfEnrollmentSettings.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import { useState } from 'preact/compat';
+import { useState } from 'react';
 import { Button, Form, InputGroup, Modal } from 'react-bootstrap';
 import { type Control, type UseFormTrigger, useWatch } from 'react-hook-form';
 
@@ -21,15 +21,15 @@ function SelfEnrollmentCode({ enrollmentCode }: { enrollmentCode: string }) {
     enrollmentCode.slice(0, 3) + '-' + enrollmentCode.slice(3, 6) + '-' + enrollmentCode.slice(6);
 
   return (
-    <div class="mb-3">
-      <label class="form-label" for="self_enrollment_code">
+    <div className="mb-3">
+      <label className="form-label" htmlFor="self_enrollment_code">
         Self-enrollment code
       </label>
       <InputGroup>
         <Form.Control
           id="self_enrollment_code"
           value={enrollmentCodeDashed}
-          style="font-family: monospace; font-size: 1.1em; letter-spacing: 0.1em;"
+          style={{ fontFamily: 'monospace', fontSize: '1.1em', letterSpacing: '0.1em' }}
           disabled
         />
         <OverlayTrigger
@@ -48,11 +48,11 @@ function SelfEnrollmentCode({ enrollmentCode }: { enrollmentCode: string }) {
               setTimeout(() => setCopied(false), 1500);
             }}
           >
-            <i class="bi bi-clipboard" />
+            <i className="bi bi-clipboard" />
           </Button>
         </OverlayTrigger>
       </InputGroup>
-      <small class="form-text text-muted">
+      <small className="form-text text-muted">
         Students can use this code to enroll in the course by entering it on the homepage or after
         clicking on any link to the course instance.
       </small>
@@ -74,8 +74,8 @@ function SelfEnrollmentLink({
   const [showConfirm, setShowConfirm] = useState(false);
   return (
     <>
-      <div class="mb-3">
-        <label class="form-label" for="self_enrollment_link">
+      <div className="mb-3">
+        <label className="form-label" htmlFor="self_enrollment_link">
           Self-enrollment link
         </label>
         <InputGroup>
@@ -96,7 +96,7 @@ function SelfEnrollmentLink({
                 setTimeout(() => setCopied(false), 1500);
               }}
             >
-              <i class="bi bi-clipboard" />
+              <i className="bi bi-clipboard" />
             </Button>
           </OverlayTrigger>
           <OverlayTrigger
@@ -111,7 +111,7 @@ function SelfEnrollmentLink({
               aria-label="Self-enrollment Link QR Code"
               onClick={() => setShowQR(true)}
             >
-              <i class="bi bi-qr-code-scan" />
+              <i className="bi bi-qr-code-scan" />
             </Button>
           </OverlayTrigger>
           {canEdit && (
@@ -127,12 +127,12 @@ function SelfEnrollmentLink({
                 aria-label="Generate new self-enrollment link"
                 onClick={() => setShowConfirm(true)}
               >
-                <i class="bi bi-arrow-repeat" />
+                <i className="bi bi-arrow-repeat" />
               </Button>
             </OverlayTrigger>
           )}
         </InputGroup>
-        <small class="form-text text-muted">
+        <small className="form-text text-muted">
           Students can use this link to immediately enroll in the course. Students can also enroll
           by entering the enrollment code on any link to the course instance.
         </small>
@@ -177,7 +177,6 @@ export function SelfEnrollmentSettings({
   trigger,
   canEdit,
   hasModernPublishing,
-  enrollmentManagementEnabled,
   studentLink,
   selfEnrollLink,
   enrollmentCode,
@@ -188,7 +187,6 @@ export function SelfEnrollmentSettings({
   trigger: UseFormTrigger<SettingsFormValues>;
   canEdit: boolean;
   hasModernPublishing: boolean;
-  enrollmentManagementEnabled: boolean;
   studentLink: string;
   selfEnrollLink: string;
   enrollmentCode: string;
@@ -213,9 +211,6 @@ export function SelfEnrollmentSettings({
     name: 'self_enrollment_enabled_before_date',
   });
 
-  const { invalid: showInEnrollPageInvalid, error: showInEnrollPageError } =
-    control.getFieldState('show_in_enroll_page');
-
   const {
     invalid: selfEnrollmentEnabledBeforeDateInvalid,
     error: selfEnrollmentEnabledBeforeDateError,
@@ -223,10 +218,10 @@ export function SelfEnrollmentSettings({
 
   return (
     <>
-      <h2 class="h4">Self-enrollment</h2>
+      <h2 className="h4">Self-enrollment</h2>
 
-      {!hasModernPublishing && enrollmentManagementEnabled ? (
-        <div class="alert alert-warning">
+      {!hasModernPublishing ? (
+        <div className="alert alert-warning">
           You are using access rules to control who can access the course instance. To change the
           self-enrollment settings, you must first{' '}
           <a href="https://prairielearn.readthedocs.io/en/latest/courseInstance/#migrating-from-allowaccess">
@@ -236,146 +231,81 @@ export function SelfEnrollmentSettings({
         </div>
       ) : null}
 
-      <div class={clsx('mb-3 form-check', !enrollmentManagementEnabled && 'd-none')}>
+      <div className="mb-3 form-check">
         <input
-          class="form-check-input"
+          className="form-check-input"
           type="checkbox"
           id="self_enrollment_enabled"
-          disabled={!canEdit || !hasModernPublishing || !enrollmentManagementEnabled}
-          {...control.register('self_enrollment_enabled', {
-            // Re-run validation on show_in_enroll_page when self-enrollment changes
-            deps: ['show_in_enroll_page'],
-          })}
+          disabled={!canEdit || !hasModernPublishing}
+          {...control.register('self_enrollment_enabled')}
         />
-        {(!canEdit || !hasModernPublishing || !enrollmentManagementEnabled) && (
+        {(!canEdit || !hasModernPublishing) && (
           <input
             type="hidden"
             name="self_enrollment_enabled"
             value={selfEnrollmentEnabled ? 'checked' : ''}
           />
         )}
-        <label class="form-check-label" for="self_enrollment_enabled">
+        <label className="form-check-label" htmlFor="self_enrollment_enabled">
           Allow self-enrollment
         </label>
-        <div class="small text-muted">
+        <div className="small text-muted">
           If not checked, students will need to be invited to this course instance.
         </div>
       </div>
 
-      <div class={clsx('mb-3 form-check', enrollmentManagementEnabled && 'd-none')}>
+      <div className={clsx('mb-3 form-check', !selfEnrollmentEnabled && 'd-none')}>
         <input
-          class={clsx('form-check-input', showInEnrollPageInvalid && 'is-invalid')}
-          type="checkbox"
-          id="show_in_enroll_page"
-          {...control.register('show_in_enroll_page', {
-            validate: (value, { self_enrollment_enabled }) => {
-              if (!self_enrollment_enabled && value) {
-                return '"Allow self-enrollment" must be checked in order to check "Show on enrollment page"';
-              }
-              return true;
-            },
-          })}
-        />
-        <label class="form-check-label" for="show_in_enroll_page">
-          Show on enrollment page
-        </label>
-        {showInEnrollPageError ? (
-          <div class="invalid-feedback">{showInEnrollPageError.message}</div>
-        ) : (
-          <div class="small text-muted">
-            If not checked, students will need a direct link to the course instance to enroll.
-          </div>
-        )}
-      </div>
-
-      <div
-        class={clsx(
-          'mb-3 form-check',
-          (!enrollmentManagementEnabled || !selfEnrollmentEnabled) && 'd-none',
-        )}
-      >
-        <input
-          class={clsx('form-check-input')}
+          className={clsx('form-check-input')}
           type="checkbox"
           id="self_enrollment_use_enrollment_code"
-          disabled={
-            !canEdit ||
-            !selfEnrollmentEnabled ||
-            !hasModernPublishing ||
-            !enrollmentManagementEnabled
-          }
+          disabled={!canEdit || !selfEnrollmentEnabled || !hasModernPublishing}
           {...control.register('self_enrollment_use_enrollment_code')}
         />
-        {(!canEdit ||
-          !selfEnrollmentEnabled ||
-          !hasModernPublishing ||
-          !enrollmentManagementEnabled) && (
+        {(!canEdit || !selfEnrollmentEnabled || !hasModernPublishing) && (
           <input
             type="hidden"
             name="self_enrollment_use_enrollment_code"
             value={selfEnrollmentUseEnrollmentCode ? 'checked' : ''}
           />
         )}
-        <label class="form-check-label" for="self_enrollment_use_enrollment_code">
+        <label className="form-check-label" htmlFor="self_enrollment_use_enrollment_code">
           Use enrollment code for self-enrollment
         </label>
-        <div class="small text-muted">
+        <div className="small text-muted">
           If not checked, any link to to anything in the course instance will allow self-enrollment.
         </div>
       </div>
 
-      <div
-        class={clsx(
-          'mb-3 form-check',
-          (!enrollmentManagementEnabled || !selfEnrollmentEnabled) && 'd-none',
-        )}
-      >
+      <div className={clsx('mb-3 form-check', !selfEnrollmentEnabled && 'd-none')}>
         <input
-          class={clsx('form-check-input')}
+          className={clsx('form-check-input')}
           type="checkbox"
           id="self_enrollment_restrict_to_institution"
-          disabled={
-            !canEdit ||
-            !selfEnrollmentEnabled ||
-            !hasModernPublishing ||
-            !enrollmentManagementEnabled
-          }
+          disabled={!canEdit || !selfEnrollmentEnabled || !hasModernPublishing}
           {...control.register('self_enrollment_restrict_to_institution')}
         />
-        {(!canEdit ||
-          !selfEnrollmentEnabled ||
-          !hasModernPublishing ||
-          !enrollmentManagementEnabled) && (
+        {(!canEdit || !selfEnrollmentEnabled || !hasModernPublishing) && (
           <input
             type="hidden"
             name="self_enrollment_restrict_to_institution"
             value={selfEnrollmentRestrictToInstitution ? 'checked' : ''}
           />
         )}
-        <label class="form-check-label" for="self_enrollment_restrict_to_institution">
+        <label className="form-check-label" htmlFor="self_enrollment_restrict_to_institution">
           Restrict self-enrollment to institution "{institution.long_name}"
         </label>
-        <div class="small text-muted">
+        <div className="small text-muted">
           If not checked, users from any institution can self-enroll.
         </div>
       </div>
 
-      <div
-        class={clsx(
-          'mb-3 form-check',
-          (!enrollmentManagementEnabled || !selfEnrollmentEnabled) && 'd-none',
-        )}
-      >
+      <div className={clsx('mb-3 form-check', !selfEnrollmentEnabled && 'd-none')}>
         <input
-          class={clsx('form-check-input')}
+          className={clsx('form-check-input')}
           type="checkbox"
           id="disable_self_enrollment_after_date"
-          disabled={
-            !canEdit ||
-            !hasModernPublishing ||
-            !enrollmentManagementEnabled ||
-            !selfEnrollmentEnabled
-          }
+          disabled={!canEdit || !hasModernPublishing || !selfEnrollmentEnabled}
           {...control.register('self_enrollment_enabled_before_date_enabled', {
             onChange: async (event) => {
               if (!event.target.checked) {
@@ -384,20 +314,17 @@ export function SelfEnrollmentSettings({
             },
           })}
         />
-        {(!canEdit ||
-          !hasModernPublishing ||
-          !enrollmentManagementEnabled ||
-          !selfEnrollmentEnabled) && (
+        {(!canEdit || !hasModernPublishing || !selfEnrollmentEnabled) && (
           <input
             type="hidden"
             name="self_enrollment_enabled_before_date_enabled"
             value={selfEnrollmentEnabledBeforeDateEnabled ? 'checked' : ''}
           />
         )}
-        <label class="form-check-label" for="disable_self_enrollment_after_date">
+        <label className="form-check-label" htmlFor="disable_self_enrollment_after_date">
           Self-enrollment cutoff date
         </label>
-        <div class="small text-muted">
+        <div className="small text-muted">
           If set, self-enrollment will be disabled after this date. We recommend setting this to the
           University-imposed deadline for students to add courses.
         </div>
@@ -405,12 +332,20 @@ export function SelfEnrollmentSettings({
         <input
           type="datetime-local"
           aria-label="Self-enrollment cutoff date"
-          class={clsx('form-control mt-2', selfEnrollmentEnabledBeforeDateInvalid && 'is-invalid')}
+          className={clsx(
+            'form-control mt-2',
+            selfEnrollmentEnabledBeforeDateInvalid && 'is-invalid',
+          )}
+          aria-invalid={selfEnrollmentEnabledBeforeDateInvalid}
+          aria-errormessage={
+            selfEnrollmentEnabledBeforeDateInvalid
+              ? 'self-enrollment-enabled-before-date-error'
+              : undefined
+          }
           disabled={
             !canEdit ||
             !selfEnrollmentEnabledBeforeDateEnabled ||
             !hasModernPublishing ||
-            !enrollmentManagementEnabled ||
             !selfEnrollmentEnabled
           }
           step="1"
@@ -426,7 +361,6 @@ export function SelfEnrollmentSettings({
         {(!canEdit ||
           !selfEnrollmentEnabledBeforeDateEnabled ||
           !hasModernPublishing ||
-          !enrollmentManagementEnabled ||
           !selfEnrollmentEnabled) && (
           <input
             type="hidden"
@@ -435,11 +369,13 @@ export function SelfEnrollmentSettings({
           />
         )}
         {selfEnrollmentEnabledBeforeDateError && (
-          <div class="invalid-feedback">{selfEnrollmentEnabledBeforeDateError.message}</div>
+          <div className="invalid-feedback" id="self-enrollment-enabled-before-date-error">
+            {selfEnrollmentEnabledBeforeDateError.message}
+          </div>
         )}
       </div>
 
-      {selfEnrollmentEnabled && enrollmentManagementEnabled && selfEnrollmentUseEnrollmentCode ? (
+      {selfEnrollmentEnabled && selfEnrollmentUseEnrollmentCode ? (
         <>
           <SelfEnrollmentCode enrollmentCode={enrollmentCode} />
           <SelfEnrollmentLink